spdlog integration + small changes

+++ winkrnl/src/vuln/drivers/basic_vuln_driver.hpp
@@ -21,7 +21,7 @@ public:
     const std::string& GetName() const;
 
 public:
-    virtual uintptr_t GetTimeStamp() const = 0;
+    virtual ULONG GetTimeStamp() const = 0;
     virtual const std::vector<uint8_t>& GetData() const = 0;
 
 public:
